requirements.txt
python-dotenv>=1.0.0
streamlit>=1.42.0
google-genai>=1.0.0
google-cloud-bigquery>=3.25.0
google-cloud-aiplatform>=1.70.0
google-auth>=2.35.0
pandas>=2.2.0
pyarrow>=18.0.0
db-dtypes>=1.3.0
duckdb>=1.1.0
openpyxl>=3.1.0
google-cloud-storage>=2.10.0
google-cloud-orchestration-airflow>=1.11.0
requests>=2.31.0
gitpython>=3.1.0
